Refactor limit to buy on credit certain buildings
Some buildings (solar power stations, universities, parks, recycles and
rocket pad launchers) cannot be payed on credit. This feature was
handled by `GROUP_*_NO_CREDIT` defines, but there was also an unused
`no_credit` field in the `struct GROUP`. This change removes the
conditional compilation and uses the boolean field to check if a
building can be purchased on credit.

 * [engine.c] Refactor `no_credit_build()` to use `no_credit` field.
   Replace the call by the new simpler function body and remove the
   function

 * [mouse.c] Refactor `no_credit_build_msg()` to not use
   `GROUP_*_NO_CREDIT` conditional compilation

 * [lintypes.c] Fix the `no_credit` field of the university group
   buildings (must be true)

 * [lintypes.h] Remove `GROUP_*_NO_CREDIT` defines from `lintypes.h`
   since they are no longer used in `engine.c` and `mouse.c`